refactor(i18n): enhance modularity for the shared locales
