+++ packages/i18n/src/packages/ui/types.d.ts
@@ -1,11 +1,11 @@
 // This file is automatically generated by i18next-cli, because it was not existing. You can edit it based on your needs: https://www.i18next.com/overview/typescript#custom-type-options
-import Resources from "./resources.ts";
+import sharedResources from "../../shared/resources.ts";
 
 declare module "i18next" {
   interface CustomTypeOptions {
     enableSelector: false;
-    defaultNS: "ui";
-    resources: Resources;
+    defaultNS: false;
+    resources: Pick<sharedResources, "ui">;
     strictKeyChecks: true;
   }
 }
