+++ packages/i18n/src/shared/resources.ts
