+++ packages/i18n/src/shared/types.d.ts
@@ -1,12 +1,11 @@
 // This file is automatically generated by i18next-cli, because it was not existing. You can edit it based on your needs: https://www.i18next.com/overview/typescript#custom-type-options
-import UiResources from "../../packages/ui/types.d.ts";
 import Resources from "./resources.ts";
 
 declare module "i18next" {
   interface CustomTypeOptions {
     enableSelector: false;
-    defaultNS: "pages";
-    resources: Resources & UiResources;
+    defaultNS: false;
+    resources: Resources;
     strictKeyChecks: true;
   }
 }
