+++ turbo.json
@@ -86,7 +86,7 @@
         "@dukkani/i18n#dashboard:types",
         "@dukkani/i18n#web:types",
         "@dukkani/i18n#storefront:types",
-        "@dukkani/i18n#ui:types"
+        "@dukkani/i18n#shared:types"
       ]
     },
     "i18n:status": {
@@ -94,43 +94,48 @@
         "@dukkani/i18n#dashboard:status",
         "@dukkani/i18n#web:status",
         "@dukkani/i18n#storefront:status",
-        "@dukkani/i18n#ui:status"
+        "@dukkani/i18n#shared:status"
       ]
     },
     "@dukkani/i18n#dashboard:types": {
-      "dependsOn": ["^@dukkani/i18n#ui:types"],
+      "dependsOn": ["^@dukkani/i18n#shared:types"],
       "inputs": ["src/apps/dashboard/**"]
     },
     "@dukkani/i18n#dashboard:status": {
-      "dependsOn": ["^@dukkani/i18n#ui:status"],
+      "dependsOn": ["^@dukkani/i18n#shared:status"],
       "inputs": ["src/apps/dashboard/**", "$TURBO_ROOT$/apps/dashboard/src/**"]
     },
     "@dukkani/i18n#web:types": {
-      "dependsOn": ["^@dukkani/i18n#ui:types"],
+      "dependsOn": ["^@dukkani/i18n#shared:types"],
       "inputs": ["src/apps/web/**"]
     },
     "@dukkani/i18n#web:status": {
-      "dependsOn": ["^@dukkani/i18n#ui:status"],
-      "inputs": ["src/apps/web/**", "$TURBO_ROOT$/apps/web/src/**"]
+      "dependsOn": ["^@dukkani/i18n#shared:status"],
+      "inputs": [
+        "src/apps/web/locales/fr/*.json",
+        "src/apps/web/i18next.config.ts",
+        "$TURBO_ROOT$/apps/web/src/**"
+      ]
     },
     "@dukkani/i18n#storefront:types": {
-      "dependsOn": ["^@dukkani/i18n#ui:types"],
+      "dependsOn": ["^@dukkani/i18n#shared:types"],
       "inputs": ["src/apps/storefront/**"]
     },
     "@dukkani/i18n#storefront:status": {
-      "dependsOn": ["^@dukkani/i18n#ui:status"],
+      "dependsOn": ["^@dukkani/i18n#shared:status"],
       "inputs": [
-        "src/apps/storefront/**",
+        "src/apps/storefront/locales/fr/*.json",
+        "src/apps/storefront/i18next.config.ts",
         "$TURBO_ROOT$/apps/storefront/src/**"
       ]
     },
-    "@dukkani/i18n#ui:types": {
-      "inputs": ["src/shared/locales/fr/ui.json"]
+    "@dukkani/i18n#shared:types": {
+      "inputs": ["src/shared/locales/fr/*.json"]
     },
-    "@dukkani/i18n#ui:status": {
+    "@dukkani/i18n#shared:status": {
       "inputs": [
-        "src/shared/locales/*/ui.json",
-        "src/shared/ui/**",
+        "src/shared/locales/fr/*.json",
+        "src/shared/i18next.config.ts",
         "$TURBO_ROOT$/packages/ui/src/**"
       ]
     },
